requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.1.0
discord.py==2.5.2
dotenv==0.9.9
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.3.0
python-dotenv==1.0.1
typing_extensions==4.12.2
wavelink==3.4.1
yarl==1.18.3
